Moved Vertex out of Octree file
